requirements.txt
python-telegram-bot==20.3
aiogram==2.25.2
PyPDF2==3.0.1
python-dotenv==1.0.0